Add basic server and simple database loaded from jsonnet
Add a basic server that can serve the `Search` LDAP function, backed by
a simple database. The database is loaded from a jsonnet file (which can
be just json if desired). Add anonymous binds, and a stub for password
binds.

The database is not case-insensitive, nor is it backed by a schema. It
is just a simple tree of `Entries`, each of which has a Distinguished
Name (DN) and a map of attribute names to slice of values. Values are
of type string only.

This is a far cry from a proper LDAP database, but serves well as a
stand-in during development.

Disable a couple of gocritic linters because they're stupid.